Split up the workflow and fix warnings
