+++ crates/holochain/src/core/workflow/countersigning_workflow/accept.rs
@@ -68,7 +68,10 @@ pub async fn accept_countersigning_request(
                 Err(KitsuneError::other("Session already exists"))
             }
             std::collections::hash_map::Entry::Vacant(entry) => {
-                tracing::debug!("Storing accepted session in the workspace for agent: {:?}", author);
+                tracing::debug!(
+                    "Storing accepted session in the workspace for agent: {:?}",
+                    author
+                );
                 entry.insert(SessionState::Accepted(request.clone()));
                 Ok(())
             }
